refactor: prepare for code migration by updating file extensions to .ts and creating interfaces
