--- src/components/controller/appLoader.js
@@ -1,11 +1,0 @@
-import Loader from './loader';
-
-class AppLoader extends Loader {
-    constructor() {
-        super(process.env.API_URL, {
-            apiKey: process.env.API_KEY,
-        });
-    }
-}
-
-export default AppLoader;
