+++ src/components/view/appView.ts
@@ -1,0 +1,96 @@
+import News from './news/news';
+import Sources from './sources/sources';
+
+interface ISource {
+  id: string;
+  name: string;
+  description: string;
+  url: string;
+  category: string;
+  language: string;
+  country: string;
+}
+
+interface ISourcesResponse {
+  status: string;
+  sources: ISource[];
+}
+
+interface ISourcesRequestParams {
+  category?: 'business' | 'entertainment' | 'general' | 'health' | 'science' | 'sports' | 'technology';
+  language?: 'ar' | 'de' | 'en' | 'es' | 'fr' | 'he' | 'it' | 'nl' | 'no' | 'pt' | 'ru' | 'sv' | 'ud' | 'zh';
+  country?:
+    | 'ae'
+    | 'ar'
+    | 'at'
+    | 'au'
+    | 'be'
+    | 'bg'
+    | 'br'
+    | 'ca'
+    | 'ch'
+    | 'cn'
+    | 'co'
+    | 'cu'
+    | 'cz'
+    | 'de'
+    | 'eg'
+    | 'fr'
+    | 'gb'
+    | 'gr'
+    | 'hk'
+    | 'hu'
+    | 'id'
+    | 'ie'
+    | 'il'
+    | 'in'
+    | 'it'
+    | 'jp'
+    | 'kr'
+    | 'lt'
+    | 'lv'
+    | 'ma'
+    | 'mx'
+    | 'my'
+    | 'ng'
+    | 'nl'
+    | 'no'
+    | 'nz'
+    | 'ph'
+    | 'pl'
+    | 'pt'
+    | 'ro'
+    | 'rs'
+    | 'ru'
+    | 'sa'
+    | 'se'
+    | 'sg'
+    | 'si'
+    | 'sk'
+    | 'th'
+    | 'tr'
+    | 'tw'
+    | 'ua'
+    | 'us'
+    | 've'
+    | 'za';
+}
+
+export class AppView {
+  constructor() {
+    this.news = new News();
+    this.sources = new Sources();
+  }
+
+  drawNews(data) {
+    const values = data?.articles ? data?.articles : [];
+    this.news.draw(values);
+  }
+
+  drawSources(data) {
+    const values = data?.sources ? data?.sources : [];
+    this.sources.draw(values);
+  }
+}
+
+export default AppView;
